Move Exceptions to dedicated namespace
