+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   instead of `Psr\Http\Message\UriInterface`
 - `Genius\Genius::getRequestFactory` returns now `Psr\Http\Message\RequestFactoryInterface` 
   instead of `Http\Message\MessageFactory`
+- Moved `Genius\GeniusException` to `Genius\Exception\GeniusException`
+- Moved `Genius\ConnectGeniusException` to `Genius\Exception\ConnectGeniusException`
+- Moved `Genius\Resources\ResourceException` to `Genius\Exception\ResourceException`
 
 ### Removed
