+++ src/Genius/Exception/ApiResponseErrorException.php
@@ -3,8 +3,6 @@ declare(strict_types=1);
 
 namespace Genius\Exception;
 
-use Genius\GeniusException;
-
 class ApiResponseErrorException extends GeniusException {
 
 }
