+++ src/Genius/Exception/ConnectGeniusException.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Genius;
+namespace Genius\Exception;
 
 final class ConnectGeniusException extends GeniusException
 {
