+++ src/Genius/Exception/ResourceException.php
@@ -1,9 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Genius\Resources;
-
-use Genius\GeniusException;
+namespace Genius\Exception;
 
 final class ResourceException extends GeniusException
 {
